Prototype design with some Boulder functionality
Implements Boulder and necessary classes to play the game in a console
interface. Also implements stub Enemy and Treasure, which will be
necessary to demonstrate one of the AC, but are not doing anything at
this point.

The Boulder functionality which I purport to have implemented is:
- Boulders are loaded from the map file and rendered in their correct
  position (in our very basic sense of a 'map', this is true).
- Can push boulders into adjacent squares by walking into from the
  opposite direction.
- Can only push boulders if there are no walls or other boulders in its
  path. Boulders otherwise cannot be walked into (behave like walls).

This hasn't been tested yet (expect JUnit architecture soon if luck is
on my side).

--

That was too much nonsense for one commit! I am very aware, but we need
something to get things rolling, and it's very difficult going from
nothing to something small.

This was originally just an exploratory model that ought to have been
tossed aside. It is only for ease of communication (and due to the early
deadline) that it finds itself on git at the present moment. I will
shooortly be whittling this down into something well documented and worth
looking at.