Major refactoring: add Pushable interface et al.
This commit introduces a lot of changes that bring the code more so in
line with what we discussed in person the other day.

DIFFERENCE: Collidable here appears as an interface rather than a
method. We check if an entity's collidable by checking

  Entity instanceof Collidable

rather than

  Entity.isCollidable().

EFFECT: This will lock us in to having to have OpenDoor and ClosedDoors
as separate entities, but we can think about whether or not this is a
good thing then (we can probably call this "state pattern" if we're
starving enough).

DIFFERENCE: ListenForMovement exists.

EFFECT: We can use this to handle how Events respond to movement Events.
We're already lazily implementing (arguably) the Observer pattern:
instances of ListenForMovement observe Cell for Movables entering or
exiting that cell.
In the future, we might want to separate out listeners like

  class ListenForEntrance
  class ListenForExit
  class ListenForSwordSwingAtopOf

so that we don't have to write nonsense 'ListenForExit' events for
things like Enemies.

I'm not willing to implement too complex and central of an Event
handling system. I think this would be a case of "Design patterns are
missing language features". Certainly, though, there is still room for
a system that can better handle events.

Maybe even

 interface CellEvent {}
 class Entrance extends CellEvent { X pullX(){} Y pullY(){} }
 class Exit extends CellEvent { X pullX(){} }
 class SwordSwing extends CellEvent { Z pullZ(){} }

And have Cell generate events as appropriate, and have e.g.
class Treasure implements ListenForCellEvent {
   void notify(Event e) {
      if (e instanceof Entrance) {
         if ((Entrance e).getWho() instanceof player) {
            treasure.pickup(...);
       }
   }
}

... I've half convinced myself. I'll think about this more later.