Observer update method now named "notifyOf"
This change is as opposed to "notify" as it was before, which I don't
believe to be common practice:
Though an Observer is "notified", it is the Subject that "notifies"
which is meant to have a method named "notify".

But this is all academic since Java keeps the word for itself.