ObserveCell now explicitly extends Entity
Should it make sense for something that isn't an Entity to observer a
cell? I would argue that only Entities should have to know about cells
at all (as well as the board). Things like PuzzleWinCondition shouldn't
touch cells, it should just listen for FloorSwitches, which in turn
listen for boulders being pushed on top of them.