Strategy methods are now static
I think static strategy methods are good.

I moved the EnemyMovementStrategy interface outside of where it's used,
probably in violation of DIP. Maybe. I've always thought of DIP as
asserting "Yeah yeah, higher level should depend on abstraction instead
of concrete implementation of lower level method ... and that
abstraction -- that interface -- should be placed with the higher level
module. That's why it's called "inversion", because the lower level
module actually has to import the higher level module to get the
interface".

But this feels too C-like. I'll think about this more in a moment.