Collidable interface replaced with Entity method
This makes things less rigid. Though, things were rigid for a reason --
I didn't want complex logic in isCollidable() about 'if this is a
boulder, i'm only collidable if my adjecent isn't collidable ... oh, or
of my adjacent's a portal -- in which case, it depends on blah'.

But, this is indeed rigid enough to stop us from implementing a tiny
state pattern in the Door class.

I don't necessarily think having Door implement the state pattern is any
good, but it's probably something that will be looked for in marking.
IMO, an open door has texture X, z-index Y, and collidability Z,
while a closed door has texture A, z-index B, and collidability C; and
they share no other behaviour -- the only thing in common is their
location -- then they're different entities!