Cells now request redraw appropriately
The 'appropriateness' when it comes to redrawing open doors is a
coincidence. When a door opens, it doesn't trigger a CellRedrawEvent,
but luckily, the player entity entering the door cell triggers the
redraw event.

We could easily change it so that Door triggers a redraw event on state
change ... I might do so.