Updated comment about how basic our JSON tests are
Specifically, they are intended to be basic here. More complicated goal
testing belongs in the parameterised TestCompositeGoal test, and as for
the non-goal loader, we've already got an adapter piping all board
creation through to it.

The basic tests are just asserting that we haven't made an error at one
end of the adapter, and accidentally fixed it at the latter. To me, it
seems like the adapter logic is faily simple, and it's unlikely that
such a mistake has been made silently.